add create script

+++ bin/dept-react-scripts.js
@@ -3,8 +3,12 @@ const { program } = require('commander')
 const { runner } = require('hygen')
 const fs = require('fs')
 const execa = require('execa')
+const { execSync } = require('child_process')
 const Logger = require('hygen/lib/logger')
 const path = require('path')
+const Case = require('case')
+const chalk = require('chalk')
+const { Confirm } = require('enquirer')
 
 const packageJson = require('../package.json')
 const defaultTemplates = path.join(__dirname, '../templates')
@@ -12,8 +16,42 @@ const defaultTemplates = path.join(__dirname, '../templates')
 program.version(packageJson.version)
 
 program
+  .command('create <name>')
+  .description('Create a new Dept React project')
+  .action(name => {
+    try {
+      const projectName = Case.snake(name)
+      execSync(`npx degit git@bitbucket.org:tamtam-nl/dtnl-dept-react-setup.git ${projectName}`)
+      const projectFilePath = `${projectName}/package.json`
+      const projectFile = fs.readFileSync(projectFilePath, 'utf-8')
+      const projectObject = JSON.parse(projectFile)
+      projectObject.name = projectName
+      projectObject.version = '1.0.0'
+      projectObject.description = 'Dept React project'
+      projectObject.contributors = []
+      delete projectObject.repository
+      fs.writeFileSync(projectFilePath, JSON.stringify(projectObject), 'utf-8')
+      execSync(`npx prettier --write ${projectFilePath}`)
+      execSync(
+        `git init && git add . && git commit -am "Initialized project with dept-react-scripts"`,
+        {
+          cwd: projectName,
+        },
+      )
+
+      console.log(
+        `Project ${chalk.green(projectName)} has succesfully been created in ${chalk.yellow(
+          `${process.cwd()}/${projectName}`,
+        )}`,
+      )
+    } catch (err) {
+      console.error(chalk.red('Something went wrong while initializing project.'))
+    }
+  })
+
+program
   .command('route <name>')
-  .description('Generate new route')
+  .description('Generate new Next.js route')
   .action(name => {
     runHygen({
       generator: name.includes('api') ? 'api' : 'page',
@@ -23,7 +61,7 @@ program
 
 program
   .command('context <name>')
-  .description('Generate context')
+  .description('Generate React context')
   .action(name => {
     runHygen({
       generator: 'context',
@@ -33,7 +71,7 @@ program
 
 program
   .command('component <name>')
-  .description('Generate a component')
+  .description('Generate a React component')
   .action(name => {
     const { path, filename } = splitPath(name)
     runHygen({
@@ -46,7 +84,16 @@ program
 program
   .command('eslint-eject')
   .description('Eject eslint config from dept-react-scripts')
-  .action(() => {
+  .action(async () => {
+    const confirm = new Confirm({
+      name: 'confirm',
+      message: `Are you sure you want to eject? This can't be undone`,
+    })
+
+    if (!(await confirm.run())) {
+      return
+    }
+
     const contents = fs.readFileSync(path.join(__dirname, '../config/eslint/index.js'))
     const eslintPath = path.join(process.cwd(), '.eslintrc.js')
     fs.writeFileSync(eslintPath, contents)
